VulnScanner/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
flask>=3.0.0
urllib3>=2.0.0
